adjusted order of execution with a gameEvent for the map
created a blend-in/out base class for enabling/disabling gameobjects. It is used for all tiles and interactables (needed some refactoring: each tile now has a script with its neccessary data - including an interactable)

the old Tile class is now only for generating. The runtime list is a List of TileBehaviour.

some cleanup

removed the bug when lost

removed a bug where the hatch was trying to spawn outside of the maze

created todos for cleanup

the maze renderSize is now 1